vpn-bot/requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
aiofiles==23.2.1
PyGithub==2.1.1
ping3==4.0.8